+++ src/main/kotlin/com/github/hotm/blockentity/NecterePortalSpawnerBlockEntity.kt
@@ -1,7 +1,6 @@
 package com.github.hotm.blockentity
 
-import com.github.hotm.HotMBlockEntities
-import com.github.hotm.HotMBlocks
+import com.github.hotm.blocks.HotMBlocks
 import com.github.hotm.world.gen.feature.NecterePortalGen
 import net.minecraft.block.BlockState
 import net.minecraft.block.Blocks
