+++ src/main/kotlin/com/github/hotm/blocks/NecterePortalBlock.kt
@@ -1,6 +1,5 @@
 package com.github.hotm.blocks
 
-import com.github.hotm.HotMBlocks
 import com.github.hotm.world.HotMDimensions
 import com.github.hotm.mixin.EntityAccessor
 import net.minecraft.block.Block
