+++ src/main/kotlin/com/github/hotm/blocks/spore/PlasseinSporeGenerator.kt
@@ -1,6 +1,6 @@
 package com.github.hotm.blocks.spore
 
-import com.github.hotm.HotMBlockTags
+import com.github.hotm.misc.HotMBlockTags
 import net.minecraft.block.BlockState
 import net.minecraft.block.Blocks
 import net.minecraft.server.world.ServerWorld
