+++ src/main/kotlin/com/github/hotm/command/HotMCommands.kt
@@ -1,4 +1,4 @@
-package com.github.hotm
+package com.github.hotm.command
 
 import com.github.hotm.command.RetrogenPortalCommand
 import net.fabricmc.fabric.api.command.v1.CommandRegistrationCallback
