+++ src/main/kotlin/com/github/hotm/config/HotMConfig.kt
@@ -1,9 +1,7 @@
 package com.github.hotm.config
 
-import com.github.hotm.HotMLog
+import com.github.hotm.misc.HotMLog
 import net.fabricmc.loader.api.FabricLoader
-import net.minecraft.util.registry.BuiltinRegistries
-import net.minecraft.world.biome.Biome
 import org.yaml.snakeyaml.Yaml
 import org.yaml.snakeyaml.constructor.Constructor
 import org.yaml.snakeyaml.introspector.PropertyUtils
