+++ src/main/kotlin/com/github/hotm/misc/HotMLog.kt
@@ -1,5 +1,6 @@
-package com.github.hotm
+package com.github.hotm.misc
 
+import com.github.hotm.HotMConstants
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
