+++ src/main/kotlin/com/github/hotm/world/gen/feature/NecterePortalGen.kt
@@ -1,9 +1,9 @@
 package com.github.hotm.world.gen.feature
 
-import com.github.hotm.HotMBlocks
+import com.github.hotm.blocks.HotMBlocks
 import com.github.hotm.config.HotMBiomesConfig
 import com.github.hotm.mixin.StructurePieceAccessor
-import com.github.hotm.world.HotMBiomeData
+import com.github.hotm.world.biome.HotMBiomeData
 import com.github.hotm.world.HotMDimensions
 import com.github.hotm.world.HotMPortalOffsets
 import net.minecraft.block.*
