+++ src/main/kotlin/com/github/hotm/world/gen/feature/TransmissionTowerFeature.kt
@@ -1,14 +1,12 @@
 package com.github.hotm.world.gen.feature
 
-import com.github.hotm.HotMProperties
+import com.github.hotm.misc.HotMProperties
 import com.mojang.serialization.Codec
 import it.unimi.dsi.fastutil.objects.Object2IntOpenHashMap
 import net.minecraft.block.BlockState
 import net.minecraft.util.math.BlockPos
 import net.minecraft.util.math.Direction
 import net.minecraft.world.ServerWorldAccess
-import net.minecraft.world.StructureWorldAccess
-import net.minecraft.world.gen.chunk.ChunkGenerator
 import net.minecraft.world.gen.feature.Feature
 import net.minecraft.world.gen.feature.util.FeatureContext
 import java.util.*
