+++ src/main/kotlin/com/github/hotm/world/gen/feature/segment/FeatureSegment.kt
@@ -1,6 +1,6 @@
 package com.github.hotm.world.gen.feature.segment
 
-import com.github.hotm.HotMRegistries
+import com.github.hotm.misc.HotMRegistries
 import com.github.hotm.util.CardinalDirection
 import com.mojang.serialization.Codec
 import net.minecraft.util.math.BlockPos
