+++ src/main/kotlin/com/github/hotm/world/gen/feature/segment/SegmentedFeature.kt
@@ -1,16 +1,14 @@
 package com.github.hotm.world.gen.feature.segment
 
-import com.github.hotm.HotMProperties
+import com.github.hotm.misc.HotMProperties
 import com.google.common.collect.Maps
 import com.google.common.collect.Queues
 import com.mojang.serialization.Codec
 import net.minecraft.util.math.BlockPos
 import net.minecraft.util.math.Direction
 import net.minecraft.world.StructureWorldAccess
-import net.minecraft.world.gen.chunk.ChunkGenerator
 import net.minecraft.world.gen.feature.Feature
 import net.minecraft.world.gen.feature.util.FeatureContext
-import java.util.*
 
 /**
  * A feature that generates in parts.
